requirements.txt
langchain>=0.0.267
langchain-openai>=0.0.2
openai>=1.3.0
python-dotenv>=1.0.0
pydantic>=2.0.0
tiktoken>=0.5.1
streamlit>=1.32.0
